+++ opencti-platform/opencti-front/src/private/components/common/stix_core_relationships/views/EntityStixCoreRelationshipsContextualViewLine.tsx
@@ -75,10 +75,6 @@ const contextualViewLineFragment = graphql`
         edges {
           node {
             id
-            entity_type
-            ... on Container {
-              representative
-            }
           }
         }
       }
