+++ opencti-platform/opencti-front/src/private/components/common/stix_core_relationships/views/indicators/EntityStixCoreRelationshipsIndicatorsContextualViewLine.tsx
@@ -72,10 +72,6 @@ const contextualViewLineFragment = graphql`
       edges {
         node {
           id
-          entity_type
-          ... on Container {
-            representative
-          }
         }
       }
     }
