+++ opencti-platform/opencti-graphql/src/database/engine.js
@@ -9,7 +9,6 @@ import {
   buildPagination,
   cursorToOffset,
   ES_INDEX_PREFIX,
-  extractEntityRepresentativeName,
   isInferredIndex,
   isNotEmptyField,
   offsetToCursor,
@@ -87,6 +86,7 @@ import { ENTITY_TYPE_SETTINGS, ENTITY_TYPE_USER } from '../schema/internalObject
 import { telemetry } from '../config/tracing';
 import { isBooleanAttribute, isDateAttribute, isDateNumericOrBooleanAttribute } from '../schema/schema-attributes';
 import { convertTypeToStixType } from './stix-converter';
+import { extractEntityRepresentativeName, extractRepresentative } from './entity-representative';
 import { ENTITY_TYPE_IDENTITY_ORGANIZATION } from '../modules/organization/organization-types';
 
 const ELK_ENGINE = 'elk';
@@ -875,6 +875,7 @@ const elBuildRelation = (type, connection) => {
     [type]: null,
     [`${type}Id`]: connection.internal_id,
     [`${type}Role`]: connection.role,
+    [`${type}Representative`]: connection.representative,
     [`${type}Name`]: connection.name,
     [`${type}Type`]: R.head(connection.types),
   };
@@ -2145,12 +2146,14 @@ const prepareRelation = (thing) => {
   connections.push({
     internal_id: from.internal_id,
     name: from.name,
+    representative: extractRepresentative(from),
     types: [from.entity_type, ...getParentTypes(from.entity_type)],
     role: thing.fromRole,
   });
   connections.push({
     internal_id: to.internal_id,
     name: to.name,
+    representative: extractRepresentative(to),
     types: [to.entity_type, ...getParentTypes(to.entity_type)],
     role: thing.toRole,
   });
@@ -2323,7 +2326,7 @@ export const elUpdateEntityConnections = async (elements) => {
   }
 };
 
-export const elUpdateConnectionsOfElement = async (documentId, documentBody) => {
+const elUpdateConnectionsOfElement = async (documentId, documentBody) => {
   const source = 'def conn = ctx._source.connections.find(c -> c.internal_id == params.id); '
     + 'for (change in params.changes.entrySet()) { conn[change.getKey()] = change.getValue() }';
   return elRawUpdateByQuery({
@@ -2355,7 +2358,7 @@ export const elUpdateElement = async (instance) => {
   // If entity with a name, must update connections
   let connectionPromise = Promise.resolve();
   if (esData.name && isStixObject(instance.entity_type)) {
-    connectionPromise = elUpdateConnectionsOfElement(instance.internal_id, { name: esData.name });
+    connectionPromise = elUpdateConnectionsOfElement(instance.internal_id, { name: esData.name, representative: esData.representative });
   }
   return Promise.all([replacePromise, connectionPromise]);
 };
