+++ opencti-platform/opencti-graphql/src/database/stix-converter.ts
@@ -1,6 +1,7 @@
 import * as R from 'ramda';
 import { v4 as uuidv4, version as uuidVersion } from 'uuid';
-import { extractEntityRepresentativeName, isEmptyField, isInferredIndex } from './utils';
+import { isEmptyField, isInferredIndex } from './utils';
+import { extractEntityRepresentativeName } from './entity-representative';
 import { FunctionalError, UnsupportedError } from '../config/errors';
 import { isBasicObject } from '../schema/stixCoreObject';
 import { isBasicRelationship } from '../schema/stixRelationship';
