+++ opencti-platform/opencti-graphql/src/domain/user.js
@@ -29,7 +29,8 @@ import {
 } from '../database/middleware-loader';
 import { delEditContext, delUserContext, notify, setEditContext } from '../database/redis';
 import { findSessionsForUsers, killUserSessions, markSessionForRefresh } from '../database/session';
-import { buildPagination, extractEntityRepresentativeName, isEmptyField, isNotEmptyField } from '../database/utils';
+import { buildPagination, isEmptyField, isNotEmptyField } from '../database/utils';
+import { extractEntityRepresentativeName } from '../database/entity-representative';
 import { publishUserAction } from '../listener/UserActionListener';
 import { ENTITY_TYPE_WORKSPACE } from '../modules/workspace/workspace-types';
 import { ABSTRACT_INTERNAL_RELATIONSHIP, OPENCTI_ADMIN_UUID } from '../schema/general';
