+++ opencti-platform/opencti-graphql/src/resolvers/stixCoreObject.js
@@ -44,8 +44,8 @@ import withCancel from '../graphql/subscriptionWrapper';
 import { connectorsForEnrichment } from '../database/repository';
 import { addOrganizationRestriction, batchObjectOrganizations, removeOrganizationRestriction } from '../domain/stix';
 import { stixCoreObjectOptions } from '../schema/stixCoreObject';
-import { extractEntityRepresentative } from "../database/utils";
 import { numberOfContainersForObject } from '../domain/container';
+import { extractRepresentative } from '../database/entity-representative';
 
 const createdByLoader = batchLoader(batchCreatedBy);
 const markingDefinitionsLoader = batchLoader(batchMarkingDefinitions);
@@ -95,7 +95,7 @@ const stixCoreObjectResolvers = {
       return 'Unknown';
     },
     toStix: (stixCoreObject, _, context) => stixLoadByIdStringify(context, context.user, stixCoreObject.id),
-    representative: (container, _, __) => extractEntityRepresentative(container),
+    representative: (container, _, __) => extractRepresentative(container),
     editContext: (stixCoreObject) => fetchEditContext(stixCoreObject.id),
     stixCoreObjectsDistribution: (stixCoreObject, args, context) => stixCoreObjectsDistributionByEntity(context, context.user, { ...args, objectId: stixCoreObject.id }),
     stixCoreRelationships: (stixCoreObject, args, context) => stixCoreRelationships(context, context.user, stixCoreObject.id, args),
